+++ libs/libc/stream/lib_rawoutstream.c
@@ -42,12 +42,12 @@
  * Name: rawoutstream_puts
  ****************************************************************************/
 
-static ssize_t rawoutstream_puts(FAR struct lib_outstream_s *self,
-                                 FAR const void *buf, size_t len)
+static int rawoutstream_puts(FAR struct lib_outstream_s *self,
+                             FAR const void *buf, int len)
 {
   FAR struct lib_rawoutstream_s *stream =
                                 (FAR struct lib_rawoutstream_s *)self;
-  ssize_t nwritten = 0;
+  int nwritten = 0;
 
   do
     {
