+++ libs/libc/stream/lib_syslogstream.c
@@ -49,8 +49,8 @@ static void syslogstream_putc(FAR struct lib_outstream_s *self, int ch)
   stream->common.nput++;
 }
 
-static ssize_t syslogstream_puts(FAR struct lib_outstream_s *self,
-                                 FAR const void *buff, size_t len)
+static int syslogstream_puts(FAR struct lib_outstream_s *self,
+                             FAR const void *buff, int len)
 {
   FAR struct lib_syslogstream_s *stream =
                                        (FAR struct lib_syslogstream_s *)self;
@@ -61,7 +61,7 @@ static ssize_t syslogstream_puts(FAR struct lib_outstream_s *self,
       return 0;
     }
 
-  syslog(stream->priority, "%.*s", (int)len, (FAR const char *)buff);
+  syslog(stream->priority, "%.*s", len, (FAR const char *)buff);
   return len;
 }
